+++ pkg/aws/v1alpha3/cluster/validate_cluster.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/blang/semver/v4"
+	kustomizev1beta2 "github.com/fluxcd/kustomize-controller/api/v1beta2"
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v6/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
@@ -14,6 +15,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	admissionv1 "k8s.io/api/admission/v1"
 	authenticationv1 "k8s.io/api/authentication/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -118,6 +120,12 @@ func (v *Validator) ValidateUpdate(request *admissionv1.AdmissionRequest) (bool,
 	}
 	if capi {
 		return true, nil
+	}
+
+	// Block v18 to v19 upgrades for gitops-managed clusters.
+	err = v.EnsureGitopsPaused(cluster, oldCluster)
+	if err != nil {
+		return false, microerror.Mask(err)
 	}
 
 	err = v.Cilium(cluster, oldCluster)
@@ -425,6 +433,39 @@ func (v *Validator) ClusterExists(obj metav1.Object) error {
 				return microerror.Maskf(notAllowedError, fmt.Sprintf("Cluster %s/%s already exists", cluster.Namespace, cluster.Name))
 			}
 		}
+	}
+	return nil
+}
+
+func (v *Validator) EnsureGitopsPaused(cluster *capi.Cluster, oldCluster *capi.Cluster) error {
+	targetRelease, err := semver.New(key.Release(cluster))
+	if err != nil {
+		return err
+	}
+
+	currentRelease, err := semver.New(key.Release(oldCluster))
+	if err != nil {
+		return err
+	}
+	if aws.IsPreCiliumRelease(currentRelease) && aws.IsCiliumRelease(targetRelease) {
+		// We are trying to upgrade from v18 to v19.
+
+		ok := key.FluxKustomizationObjectKey(cluster)
+		if ok != nil {
+			kust := kustomizev1beta2.Kustomization{}
+			err = v.k8sClient.CtrlClient().Get(context.Background(), *ok, &kust)
+			if errors.IsNotFound(err) {
+				// Labels are present but Kustomization was not found. Might be running on customer infra. Don't want to block upgrade.
+				return nil
+			} else if err != nil {
+				return err
+			}
+
+			if !kust.Spec.Suspend {
+				return microerror.Maskf(notAllowedError, fmt.Sprintf("Cluster %s/%s is managed by gitops but Kustomization %s/%s is not suspended", cluster.Namespace, cluster.Name, ok.Namespace, ok.Name))
+			}
+		}
+
 	}
 	return nil
 }