+++ pkg/mutator/handler.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"time"
 
@@ -45,7 +45,7 @@ func Handler(mutator Mutator) http.HandlerFunc {
 			return
 		}
 
-		data, err := ioutil.ReadAll(request.Body)
+		data, err := io.ReadAll(request.Body)
 		if err != nil {
 			mutator.Log("level", "error", "message", "unable to read request")
 			metrics.InternalError.WithLabelValues("mutating", mutator.Resource()).Inc()